% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{assign_auto_labels}
\alias{assign_auto_labels}
\title{Assign the labels to free parameters in MxModel}
\usage{
assign_auto_labels(M)
}
\arguments{
\item{M}{\code{MxModel} object}
}
\value{
\code{MxModel} object.
}
\description{
Assign the labels to free parameters in MxModel
}
\note{
The labels are assigned in the form "matrix_row_col"
or "model_matrix_row_col". In symmetric matrices, the labels
are the same upper and lower triangle.
}
