% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{def_ci}
\alias{def_ci}
\alias{get_ci}
\title{Define confidence intervals.}
\usage{
def_ci(model, ci = character(0), ...)

get_ci(model, ci = output_tables(model))
}
\arguments{
\item{model}{\code{MxModel} object}

\item{ci}{the character vector pointing out the parameters to
compute the intervals for, see \code{\link[OpenMx]{mxCI}}.}

\item{...}{the arguments passed to mxCI.}
}
\value{
\code{MxModel}
}
\description{
The function to define and extract confidence intervals. This information
is stored in the \code{options} slot of an \code{MxModel}.
}
\note{
These functions aim to facilitate the assembly of parameters and their CIs.
\code{get_ci} is used by \code{\link{get_output_tables}}.
}
\examples{

}
