% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_descriptives.R
\name{get_cross_trait_cors}
\alias{get_cross_trait_cors}
\title{Compute the cross-twin cross-trait correlations}
\usage{
get_cross_trait_cors(dat, zyg, vars, compact = FALSE)
}
\arguments{
\item{dat}{the data.frame with the variables zyg and paste(vars, rep(1:2,
each = nv))}

\item{zyg}{the name of the zygosity variable (character string)}

\item{vars}{the names of the variables to compute the descriptives for.
If vars = c('X', 'Y', 'Z'), the function looks for
c('X1', 'Y1', 'Z1', 'X2', 'Y2', 'Z2')}

\item{compact}{return the MZ and DZ correlations in one table, MZ
correlations on the upper triangle and DZ correlations on
the lower triangle. Works only when the zygosity variable
has exactly two levels}
}
\description{
Compute the phenotypic correlations across twins and across traits,
for MZ and DZ twins. Can be used for any data on groups.
}
