% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_models.R
\name{crosslag_ace_ade}
\alias{crosslag_ace_ade}
\alias{cross_lag_ace}
\alias{cross_lag_ade}
\title{Define a cross-lag ACE model}
\usage{
cross_lag_ace(
  data,
  zyg = character(0),
  definition,
  data_type = "raw",
  sep = getOption("TwinAnalysis.sep")
)

cross_lag_ade(
  data,
  zyg = character(0),
  definition,
  data_type = "raw",
  sep = getOption("TwinAnalysis.sep")
)
}
\arguments{
\item{data}{either \code{data.frame} (for raw data) or \code{list} for
covariation/correlation input. See Note.}

\item{zyg}{name of the variable that labels zygosity in the
\code{data.frame}.}

\item{definition}{a \code{list} that describes measurement occasions.
See Note.}

\item{data_type}{type of the data (raw, cov or cor). See Note.}

\item{sep}{separator between the name of the phenotype and the label of a
twin. Default is ''.}
}
\value{
One unfitted \code{mxModel}.
}
\description{
The function defines a cross-lag ACE model for arbitrary number of
measurement occasions and arbitrary number of phenotypes within each
measurement occasion.
}
\note{
The function accepts two forms of the data: \code{data.frame}
(\code{data_type = 'raw'}) or \code{list} of covariance/correlation matrices
(\code{data_type = 'cov'}, \code{data_type = 'cor'}).

When \code{data} is a \code{data.frame}, \code{zyg} is expected to point at
the variable in \code{data} that defines zygosity groups . Zygosity
variable MUST be a factor with two labels: 'MZ' and 'DZ'.

When \code{data} is a list of covariance/correlation matrices, it must
include two named elements, 'MZ' and 'DZ'. These elements must be the lists
with following elements: \code{observed} (covariance/correlation matrix),
\code{means} (numeric vector of observed means, optional) and \code{numObs}
(number of observations).

By default, it is expected that phenotypic trait X is labeled as 'X1' in twin
1 and 'X2' in twin 2.

\code{definition} is a list of character vectors. The first character vector
includes the variables from the first measurement occasion, the second - the
variables from the second measurement occasion, etc. In the model, the
variables from the same measurement occasion are assumed to correlate.
The variables from one measurement occasion are assumed to predict the
variables from the next measurement occasion. Refer to the vignette on
genetic cross-lag for more information.

Output tables: "Variance components": proportion of variance explained by
A/C/D/E, per variable, proportion of variance unaccounted for by preceding
measurements (specific variance); "Raw variance": total variance and variance
of A/C/D/E, per variable; "Raw paths": unstandardized regression and
covariance paths for A/C/D/E; "Standardized paths": standardized regression
and covariance paths for A/C/D/E; "Phenotypic paths": total (phenotypic)
regression and covariance paths, unstandardized and standardized, and
proportion of each path explained by A/C/D/E.
}
