% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{def_stand_params}
\alias{def_stand_params}
\title{Define standardized parameters for the RAM model.}
\usage{
def_stand_params(model)
}
\arguments{
\item{model}{\code{MxModel} object of type RAM.}
}
\value{
\code{MxModel} object.
}
\description{
The function adds the algebra defining the standardized parameters of a
RAM model and a fancy table with raw and standardized estimates.
}
\note{
The standardization is based on the variances from the expected covariation
matrix of the model. One-way paths are standardized as
\eqn{b_XY * \sqrt{V_Y}/\sqrt{V_X}} (CHECK!!!). Two-way paths are
standardized as \eqn{Cov_XY / (\sqrt{V_X}\sqrt{V_Y})}.

The function defines two output tables: \code{Parameter_table} includes
one-way and two-way paths and \code{Means} include the means. The user can
identify the confidence intervals for the whole table or for its part,
see examples for \code{\link{def_ci}}.
}
