% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_descriptives.R
\name{get_fisher_z}
\alias{get_fisher_z}
\title{Perform the Fisher test for two correlations}
\usage{
get_fisher_z(dat, zyg, vars, compare = character(0))
}
\arguments{
\item{dat}{the data.frame with the variables zyg and paste(vars, rep(1:2,
each = nv))}

\item{zyg}{the name of the zygosity variable (character string),
must have exactly two levels}

\item{vars}{the names of the variables to perform the test.
If vars = c('X', 'Y', 'Z'), the function looks for
c('X1', 'Y1', 'Z1', 'X2', 'Y2', 'Z2')}

\item{compare}{the labels of groups to compare, must be either character(0)
or length of 2}
}
\description{
The Fisher r-to-z transformation implied to compare the correlations
in two groups
}
