% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{fit_stats}
\alias{fit_stats}
\title{Compute fit statistics.}
\usage{
fit_stats(model, nested_models = NULL)
}
\arguments{
\item{model}{\code{MxModel} with the reference models included.}

\item{nested_models}{the list of nested models (\code{MxModel}s) to be
compared against the target model.}
}
\value{
\code{data.frame}
}
\description{
This function assemble the model parameters and fit statistics into
a table. It also compares nested models.
}
\note{
The function uses \code{summary.MxModel} and \code{mxCompare} to assembe
the following parameters and statistics:
\itemize{
   \item{ep}~--- number of estimated parameters;
   \item{minus2LL}~--- \eqn{-2*log-likelihood};
   \item{df}~--- degrees of freedom;
   \item{AIC}~--- Akaike Information Criterion;
   \item{BIC}~--- Bayesian Information Criterion;
   \item{CFI}~--- Comparative Fit Index;
   \item{TLI}~--- Tucker-Lewis Index;
   \item{RMSEA}~--- Root Mean Square Error of Approximation;
   \item{diffLL}~--- log-likelihood difference between two models,
      chi-squared;
   \item{diffdf}~--- the difference in degrees of freedom between
      two models, chi-square df;
   \item{p}~--- p-value of the chi-square test.
}

The references will be added.

The nested models are compared against the target model by means of the
chi-square test (the same way the target model is compared against the
saturated model).

This function does not perform any computation, it is a wrapper for the
\code{OpenMx} functions aiming to simplify the preparation of the output
from SEM.
}
