% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{get_boot_ci}
\alias{get_boot_ci}
\title{Extract the bootstrapped confidence interval}
\usage{
get_boot_ci(
  model_boot,
  table_name = character(0),
  interval = c(0.025, 0.975),
  cache_file = character(0),
  force = FALSE
)
}
\arguments{
\item{model_boot}{bootstrapped \code{MxModel}}

\item{table_name}{a character vector of the names of \code{MxAlgebra} or
\code{MxMatrix} in the model for which the CI is estimated.
If this is \code{character(0)}, it will try to extract the
names of the output tables from the model using
\code{output_tables()} (a character vector stored in
\code{@output$output_tables})}

\item{interval}{the interval's quantile (default is 95% CI)}

\item{cache_file}{a string indicating the file where the parameter tables
will be stored (.rds)}

\item{force}{force the computation regardless of whether the cache file
exists or not}
}
\value{
a table in the form of \code{mlth.data.frame}
}
\description{
If the bootstrap has been applied to the model, this function helps to
extract the bootstrapped confidence interval and present it in a table
combined with the parameters' estimates. Since the extraction of CI for the
computed parameters can take long time, it also allows saving the tables
to cache. If the cache file already exists, it will load the tables from it
}
