% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{def_output_tables}
\alias{def_output_tables}
\alias{output_tables<-}
\alias{get_output_tables}
\alias{output_tables}
\title{Define and extract \code{MxModel} output tables.}
\usage{
def_output_tables(model, output_tables = character(0))

output_tables(model) <- value

get_output_tables(model, tables = character(0), CIs = TRUE)

output_tables(model)
}
\arguments{
\item{model}{\code{MxModel} object.}

\item{output_tables, value}{character vector, the names of the output tables
defined inside the model as \code{MxAlgebra}.}
}
\value{
\code{def_output_tables} returns an \code{MxModel}
        (\code{`output_tables<-`} is an alias).

        \code{output_table} returns \code{character} vector of the names of
        the output tables.

        \code{get_output_tables} returns the list of the evaluated output
        tables. If the model output includes confidence intervals, it append
        the intervals. Each table is either \code{data.frame} or
        \code{mlth.data.frame}.
}
\description{
The functions to define and extract the output tables for \code{MxModels}.
The names of the output tables are stored as the \code{output_tables}
option of an \code{MxModel} object.
}
