% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_misc.R
\name{process_twin_data}
\alias{process_twin_data}
\title{Process twin data into MxData objects}
\usage{
process_twin_data(data, selvars, data_type = "raw", zyg = character(0))
}
\arguments{
\item{data}{\code{data.frame} or \code{list}}

\item{data_type}{'raw', 'cov' or 'cor', as required by \code{type} parameter
of \code{mxData}}

\item{zyg}{factor used to split the data frame when data_type = 'raw'}
}
\value{
A list of \code{MxData} objects
}
\description{
Process twin data into MxData objects to add to MxModel.
The function accepts raw data and covariance matrices.
}
\details{
When data_type = 'raw' (default), the data must be a data frame or matrix
with raw data that will be split upon factor \code{zyg} and passed to
\code{mxData}. When data_type = 'cov' or 'cor', the data  must be a (named)
list of arguments passed to \code{mxData}, including \code{observed} -
covariation or correlation matrix, \code{numObs} - number of observations,
\code{means} - vector of means. The \code{type} argument is not required as
it is added from \code{data_type}. Refer to \code{?mxData} for details on
\code{mxData} arguments.
}
