% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{ref_models}
\alias{ref_models}
\alias{get_ref_models}
\title{Define saturated and independence models}
\usage{
ref_models(model, ref_models = NULL, run = TRUE, ...)

get_ref_models(model)
}
\arguments{
\item{model}{\code{MxModel} object}

\item{ref_models}{a named list of length two of \code{MxModels} (Saturated and Independence).
If \code{NULL}, \code{mxRefModels} is used to define reference models.}

\item{run}{should we fit the reference models? Default is \code{TRUE}.}

\item{...}{arguments passed to \code{mxRefModels}.}
}
\value{
the original model with two reference models appended in the \code{option} slot.
}
\description{
A function to define and run two reference models: Saturated and Independence.
The reference models are stored in the \code{options} slot of the target model.
}
\note{
If you provide the list with two reference models, the models have to be stored as \code{'Saturated'}
and \code{'Independence'} object within this list. Otherwise they will be thrown away.

The reference models summarize the data, so don't forget to update the reference models
if you modify the data.
}
