% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{Run bootstrap on an MxModel}
\usage{
run_bootstrap(
  model,
  cache_file = character(0),
  N = 500,
  force = FALSE,
  by_chunk = TRUE,
  chunk_size = 100
)
}
\arguments{
\item{model}{MxModel}

\item{cache_file}{a character vector indicating the cache file (.rds)}

\item{N}{number of bootstrap samples}

\item{force}{force bootstrap regardless of whether the cache file exists}

\item{by_chunk}{run the bootstrap in chunks}

\item{chunk_size}{the size of one chunk}
}
\value{
MxModel
}
\description{
It allows running boostrap and saving it to a cache file. If the cache file
already exists, it will load the bootstrapped model from cache. It can run
in chunks giving an update when each chunk is finished. It measures the time
needed to process one chunk and gives an estimation of the total execution
time.
}
\note{
The bootstrap is only supported for the models with the real (raw) data
}
