% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_misc.R
\name{twin_ref_models}
\alias{twin_ref_models}
\title{Define twin reference models}
\usage{
twin_ref_models(model, run = FALSE, ...)
}
\arguments{
\item{model}{a twin model, MxModel object}

\item{run}{whether the reference models are to be fit, default is TRUE}

\item{...}{the parameters passed to mxRun}
}
\value{

}
\description{
Define and fit saturated and indepencence models corresponding to a twin
model. The twin model has to be fit as the function takes its estimates
of covariances and means to compute starting values. The twin model has
to include MZ and DZ sub-models.
}
\examples{

}
