% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_descriptives.R
\name{get_univ_descriptives}
\alias{get_univ_descriptives}
\title{Compute the univariate twin descriptives}
\usage{
get_univ_descriptives(dat, zyg, vars)
}
\arguments{
\item{dat}{the data.frame with the variables zyg and paste(vars, rep(1:2, each = nv))}

\item{zyg}{the name of the zygosity variable (character string)}

\item{vars}{the names of the variables to compute the descriptives for.
If vars = c('X', 'Y', 'Z'), the function looks for
c('X1', 'Y1', 'Z1', 'X2', 'Y2', 'Z2')}
}
\description{
Compute number of valid observations, means, SDs and cross-twin correlations
for MZ and DZ twins. Can be used for any data on groups.
}
\details{

}
\examples{

}
