% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_models.R
\name{univ_multiv_ace_ade}
\alias{univ_multiv_ace_ade}
\alias{univariate_ace}
\alias{univariate_ade}
\alias{multivariate_ace}
\alias{multivariate_ade}
\title{Define univariate and multivariate ACE and ADE models}
\usage{
univariate_ace(
  data,
  zyg = character(0),
  ph,
  data_type = "raw",
  sep = getOption("TwinAnalysis.sep")
)

univariate_ade(
  data,
  zyg = character(0),
  ph,
  data_type = "raw",
  sep = getOption("TwinAnalysis.sep")
)

multivariate_ace(
  data,
  zyg = character(0),
  ph,
  data_type = "raw",
  sep = getOption("TwinAnalysis.sep")
)

multivariate_ade(
  data,
  zyg = character(0),
  ph,
  data_type = "raw",
  sep = getOption("TwinAnalysis.sep")
)
}
\arguments{
\item{data}{either \code{data.frame} (for raw data) or \code{list} (for
covariance/correlation input). See Note.}

\item{zyg}{name of the variable that labels zygosity in the
\code{data.frame}.}

\item{ph}{name of the phenotype, a single character value for univariate
ACE/ADE or a character vector for multivariate ACE/ADE.}

\item{data_type}{type of the data (raw, cov or cor). See Note.}

\item{sep}{separator between the name of the phenotype and the label of a
twin. Default is "" (empty string).}
}
\value{
One unfitted \code{mxModel}.
}
\description{
The functions define univariate and multivariate (correlated factors) ACE
and ADE models.
}
\note{
The function accepts two forms of the data: \code{data.frame}
(\code{data_type = 'raw'}) or \code{list} of covariance/correlation matrices
(\code{data_type = 'cov'}, \code{data_type = 'cor'}).

When \code{data} is a \code{data.frame}, \code{zyg} is expected to point at
the variable in \code{data} that defines zygosity groups . Zygosity
variable MUST be a factor with two labels: 'MZ' and 'DZ'.

When \code{data} is a list of covariance/correlation matrices, it must
include two named elements, 'MZ' and 'DZ'. These elements must be the lists
with following elements: \code{observed} (covariance/correlation matrix),
\code{means} (numeric vector of observed means, optional) and \code{numObs}
(number of observations).

By default, it is expected that phenotypic trait X is labeled as 'X1' in twin
1 and 'X2' in twin 2.

Output tables (univariate ACE/ADE): "Variance components": proportion of
total variance explained by A/C/D/E; "Raw variance": total variance,
variance of A/C/D/E.

Output tables (multivariate ACE/ADE): "Variance components": proportion of
total variance explained by A/C/D/E, per phenotype; "Covariation
components": total covariance, covariance of A/C/D/E, proportion of
covariation explained by A/C/D/E; "All correlations": correlations
for A/C/D/E and total (phenotypic) correlation; "A/C/D/E/Total
correlations": correlation table for A/C/D/E or phenotypic correlation
(not returned by \code{get_output_tables}).
}
