% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_misc.R
\name{def_ref_models}
\alias{def_ref_models}
\alias{ref_models<-}
\alias{ref_models}
\alias{run_ref_models}
\title{Define saturared and independence models.}
\usage{
def_ref_models(model, ref_models = NULL, ...)

ref_models(model) <- value

ref_models(model)

run_ref_models(model, ...)
}
\arguments{
\item{model}{\code{MxModel} object}

\item{ref_models}{the list of length two of \code{MxModel}s
(saturated and independence) or \code{NULL}.}

\item{...}{arguments passed to \code{mxRefModels} or \code{mxRun}
(e.g., \code{run = TRUE}).}
}
\value{
\code{def_ref_models} returns the \code{MxModel}
        (\code{`ref_models<-`} is an alias)).

        \code{ref_models} returns the list of \code{MxModel}s.

        \code{run_ref_models} runs the parameter estimation for the
        reference models and return the \code{MxModel}.
}
\description{
The functions to define, fit and store the reference models.
The reference models are stored in the \code{options} slot of the target
model.
}
\note{
The idea is to store the reference models together with the original model,
so the function \code{fit_stats} could address them and compute fit
statistics. The user can define the reference models and pass them to the
function (as a list of two \code{MxModels}). When the argument
\code{ref_models} is empty or there is no Saturated or Independence model
in it, the function makes use of \code{mxRefModels}.

Don't forget to re-define the reference models when you modify the
model's data.
}
